New config format
The most important change here is the introduction of a new config format.
The config now lives in a file called 'modthegungeon.conf'.
It is now appropriately located on OS X (in Library/Application Support instead of the home directory).

Files.xml was merged into ETGMod.Installer.csproj because MonoDevelop didn't quite like this setup.

All config values will, of course, be reset with this version.